feat: "ft_printf" now prints pointers correctly.
Paco validated, pending strict mode

+++ ft_printf/ft_putnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: judelgad <judelgad@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/23 17:45:19 by judelgad          #+#    #+#             */
-/*   Updated: 2023/12/26 21:35:30 by judelgad         ###   ########.fr       */
+/*   Updated: 2023/12/27 00:45:29 by judelgad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 #include <unistd.h>
 #include <stdint.h>
 
-int	ft_putnbr_base(intptr_t n, const char *base, int *count)
+int	ft_putnbr_base(long long n, const char *base, int *count)
 {
 	size_t	base_len;
 
 	base_len = ft_strlen(base);
-	if(n < 0)
+	if (n < 0)
 	{
 		ft_putchar('-', count);
 		n *= -1;
